David

+++ core/src/com/mygdx/auber/entities/Infiltrator.java
@@ -13,9 +13,10 @@ import com.mygdx.auber.Screens.PlayScreen;
 
 public class Infiltrator extends NPC{
     public boolean isDestroying = false;
+    public float destroyingTime;
     public double timeToWait = Math.random() * 15;
 
-    private float timeInvisibleStart;
+    private float timeInvisible;
     private boolean isInvisible = false;
 
     private boolean isMoving = true;
@@ -38,48 +39,55 @@ public class Infiltrator extends NPC{
         return;
     }
 
+    //TODO: Fix infiltrators stopping destroying key systems
     /**
      * Step needs to be called in the update method, makes the NPC move and check if it has reached its next node
      */
     public void step(float delta) {
-        this.moveNPC();
+        this.moveNPC(); //Moves the npc and sets their scale
 
         if(isDestroying)
         {
-            if(Vector2.dst(Player.x, Player.y, this.getX(), this.getY()) < 50)
+            KeySystem keySystem = KeySystemManager.getClosestKeySystem(this.getX(), this.getY());
+
+            if(keySystem.isDestroyed())
             {
+                this.isDestroying = false;
+                this.pathQueue.clear();
+                this.setGoal(MapGraph.getRandomNode());
+            }
+
+            if(Vector2.dst(Player.x, Player.y, this.getX(), this.getY()) < 75)
+            {
+                keySystem.stopDestroy();
                 this.useAbility();
                 this.isDestroying = false;
             }
-            else
-            {
-                this.elapsedTime += delta;
-            }
-            return;
-        }
+        } //If isDestroying, if the distance to the player is less than 50, use ability and stop destroying, else keep adding time
 
         if(isInvisible)
         {
-            if(System.currentTimeMillis() - timeInvisibleStart  > 10 * 100)
+            timeInvisible += delta;
+            if(timeInvisible  > 10)
             {
                 this.isInvisible = false;
             }
         }
         else
         {
+            timeInvisible = 0;
             this.setAlpha(.99f);
-        }
+        } //If isInvisible, keep adding time to timeInvisible, and if its longer than 10 seconds set isInvisible to false. If not timeInvisible, set the alpha to 1 and time to 0
 
         this.elapsedTime += delta;
-        this.checkCollision();
+        this.checkCollision(); //Add elapsed time and check collisions
 
         if((this.elapsedTime >= timeToWait) && this.pathQueue.isEmpty()) {
             this.elapsedTime = 0;
             reachDestination();
-        }
+        } //If there is no queue and elapsed time is greater than time to wait, reach destination
     }
 
-    //TODO: Redo this function
     /**
      * Called when the path queue is empty
      */
@@ -90,18 +98,26 @@ public class Infiltrator extends NPC{
         this.velocity.y = 0;
         timeToWait = Math.random() * 15;
 
-        if(Math.random() > .95f && !this.isDestroying && !this.isInvisible) // 1/10 chance of infiltrator deciding to destroy a keysystem
+        if(Math.random() > .95f && !this.isDestroying && !this.isInvisible && KeySystemManager.safeKeySystemsCount() != 0) // 1/10 chance of infiltrator deciding to destroy a keysystem
         {
             this.destroyKeySystem();
             return;
-        }
+        } //If not invisible or currently destroying a key system, random chance to go destroying a key system
 
         if(pathQueue.size == 0 && GraphCreator.keySystemsNodes.contains(this.previousNode, false))
         {
             this.isDestroying = true;
-            //KeySystem.startDestroy();
-            return;
-        }
+            KeySystem keySystem = KeySystemManager.getClosestKeySystem(this.getX(), this.getY());
+            if(keySystem == null)
+            {
+                setGoal(MapGraph.getRandomNode());
+                return;
+            }
+            if(keySystem.isSafe())
+            {
+                keySystem.startDestroy();
+            }
+        } //If no queue, and the last node in queue was a key systems node, start destroying
 
         Node newGoal;
         do {
@@ -109,7 +125,7 @@ public class Infiltrator extends NPC{
         }while(newGoal == previousNode);
         {
             setGoal(newGoal);
-        }
+        } //Set a new goal node and start moving towards it
 
     }
 
@@ -120,7 +136,7 @@ public class Infiltrator extends NPC{
     {
         if(this.pathQueue.size > 0){
             Node targetNode = this.pathQueue.first();
-            if(Vector2.dst(this.getX(),this.getY(),targetNode.x,targetNode.y) <= 10)
+            if(Vector2.dst(this.getX(),this.getY(),targetNode.x,targetNode.y) <= 15)
             {
                 reachNextNode(); //If the sprite is within 5 pixels of the node, it has reached the node
             }
@@ -132,10 +148,18 @@ public class Infiltrator extends NPC{
      */
     public void destroyKeySystem()
     {
-        System.out.println("Infiltrator moving to destroy");
         this.pathQueue.clear();
-        Node keySystem = GraphCreator.keySystemsNodes.random();
-        this.setGoal(keySystem);
+        Node keySystemNode = GraphCreator.keySystemsNodes.random();
+        KeySystem keySystem = KeySystemManager.getClosestKeySystem(keySystemNode.x, keySystemNode.y);
+
+        if((keySystem.isDestroyed() || keySystem.isBeingDestroyed()) && KeySystemManager.safeKeySystemsCount() != 0)
+        {
+            destroyKeySystem();
+        }
+        else
+        {
+            this.setGoal(keySystemNode);
+        } //If Key system is being destroyed or is already destroyed, select a new key system
     }
 
     /**
@@ -169,7 +193,7 @@ public class Infiltrator extends NPC{
     {
         this.isInvisible = true;
         this.isDestroying = false;
-        this.timeInvisibleStart = System.currentTimeMillis();
+        this.timeInvisible = 0;
         this.setAlpha(0.05f);
     }
 
@@ -187,8 +211,9 @@ public class Infiltrator extends NPC{
      */
     private void stopAuberHealing()
     {
+        System.out.println("Stopped healing");
         Player.canHeal = false;
-        Player.healStopTime = System.currentTimeMillis();
+        Player.healStopTime = 0;
     }
 
     /**
